v15/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.23.0
eth-account>=0.9.0
requests>=2.31.0
python-multipart>=0.0.6
